requirements.txt
--index-url https://gitlab.com/api/v4/projects/70323435/packages/pypi/simple
model_helpers==0.1.0
azure-core==1.34.0
azure-storage-blob==12.25.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
databricks-sdk==0.53.0
huggingface-hub==0.31.4
joblib==1.5.0
matplotlib==3.10.3
mlflow==2.22.0
mlflow-skinny==2.22.0
numpy==2.2.6
pandas==2.2.3
pillow==11.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
sympy==1.14.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.2
fairscale==0.4.13
